refactor(server): extract the A2A surface to server/a2a.py (ADR 0023 phase 2)
Second backend extraction. Moves the A2A-specific logic out of the composition
root into server/a2a.py:
- agent-card building (_build_agent_card_proto, _a2a_card_url)
- skill declarations (_SKILL_SPECS, _agent_skills, structured_skill_schema)
- per-turn telemetry writer (_record_a2a_telemetry)
- executor terminal hook (_a2a_terminal)

server/__init__.py re-exports every name (the re-export sits after agent_name /
_event_bus / _bundle_root are defined, so a2a.py importing those three from
`server` is not a cycle). `server.<symbol>` and `from server import X` are
unchanged; _main's a2a-sdk route wiring still calls them and stays put (it
shrinks in phase 3). server/__init__.py drops ~194 lines.

The only contract that moved: the test that *monkeypatches* the skill list now
patches `server.a2a._SKILL_SPECS` (its new home) — editing the source list, the
fork path, is unaffected.

Verified: 1000 tests green; live smoke against a throwaway instance booted via
`python -m server` — agent card built by the extracted _build_agent_card_proto
(right version + 4 extensions) and a full A2A 1.0 SendMessage round-trip
(TASK_STATE_COMPLETED + cost-v1), which runs the extracted terminal hook.
